Obtain TestEnvironment on explict request only.
The submit module interface has an additional '--test', flag, that
allows the user to explicitly request a TestEnvironment. This
environment has a fake scheduler that simulates the generation and
submission of a job submission script. This enables the user to test
their submission flow and fix errors before testing in the actual
submission environment.